+++ resources/views/layouts/auth/card.blade.php
@@ -22,6 +22,7 @@
             </div>
 
             <x-toast-container />
+            <x-cookie-consent-banner />
         </div>
     </body>
 </html>