+++ resources/views/layouts/auth/split.blade.php
@@ -39,6 +39,7 @@
             </div>
 
             <x-toast-container />
+            <x-cookie-consent-banner />
         </div>
     </body>
 </html>